requirements.txt
asgiref==3.5.2
Django==3.1.7
djangorestframework==3.12.4
pytz==2022.1
sqlparse==0.4.2
